Resolved requested changes
Updated README.md with examples. Added service subdirectory. Deleted get_and_parse_transcripts_and_comemnts.
