+++ service/comment_downloader.py
@@ -1,4 +1,4 @@
- #!/usr/bin/env python
+#!/usr/bin/env python
 from __future__ import print_function
 
 import os
